renamed pattern.py to interactions.py
